Closes #1
This is more a re-write of your dependency checking than a patch, unfortunately. On the other hand, it works for large numbers of files now, while I could not find a circumstance where it did before.

I'm not much of a Java guy either, so much of this actually comes from a colleague, including the use of the DAG library. It's open source under a compatible license, so it should all be cool.

Understand if you don't want to accept this much of a change, but I've tested it manually under a series of scenarios and it seems to work fine. Obviously automated tests would be better, but I've not found the time.